# api/openapi.json
{
  "components": {
    "schemas": {
...
      "EnrollRequest": {
        "properties": {
          "password": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "token": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "password"
        ],
        "title": "EnrollRequest",
        "type": "object"
      },
      "EnrollResponse": {
        "description": "Returned exactly once. The secret and codes are never retrievable again.",
        "properties": {
          "confirm_token": {
            "title": "Confirm Token",
            "type": "string"
          },
          "hint": {
            "title": "Hint",
            "type": "string"
          },
          "otpauth_uri": {
            "title": "Otpauth Uri",
            "type": "string"
          },
          "recovery_codes": {
            "items": {
              "type": "string"
            },
            "title": "Recovery Codes",
            "type": "array"
          },
          "secret": {
            "title": "Secret",
            "type": "string"
          }
        },
        "required": [
          "secret",
          "otpauth_uri",
          "recovery_codes",
          "hint",
          "confirm_token"
        ],
        "title": "EnrollResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Two surfaces sharing one deployment: a public capture surface (`/r/{slug}`) and an authenticated admin API (`/api/v1/*`). See docs/API.md.",
    "summary": "Visitor intelligence platform",
    "title": "Tracelet",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/enroll": {
      "post": {
        "description": "Sets the first password and returns the TOTP secret plus ten recovery codes. **Returned exactly once** -- neither is retrievable afterwards. The account becomes active only after /totp/confirm proves the authenticator works.",
        "operationId": "enroll_api_v1_auth_enroll_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete enrolment with a one-time token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "Verifies the password and returns a short-lived MFA token. **Not** a session: TOTP is mandatory, so no route is reachable until step 2 completes. Failures are indistinguishable for wrong password, unknown account, disabled account and incomplete enrolment.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Step 1 of 2: password",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/me/preferences": {
      "patch": {
        "description": "Display only. The theme is semi-dark unless changed (F9.AC16); the timezone is how timestamps are shown to you, not how analytics are bucketed (ADR-0016). Not audited: it changes nothing anyone else sees.",
        "operationId": "update_preferences_api_v1_auth_me_preferences_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change your own theme or display timezone",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/mfa": {
      "post": {
        "description": "Validates the TOTP code and issues the session cookie. A code at or below the last accepted time step is rejected as reused, so an observed code cannot be replayed within its window.",
        "operationId": "mfa_api_v1_auth_mfa_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Step 2 of 2: authenticator code",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/password": {
      "post": {
        "description": "Requires the current password. Revokes every OTHER session on success, because changing a password is the action a victim takes believing it locks an attacker out.",
        "operationId": "change_own_password_api_v1_auth_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change your own password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/recovery-code": {
      "post": {
        "description": "Bypasses both password and TOTP -- that is what a recovery code is for. Single-use. The remaining count is returned in the X-Recovery-Remaining header.",
        "operationId": "use_recovery_code_api_v1_auth_recovery_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoveryCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sign in with a recovery code",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/reset/confirm": {
      "post": {
        "operationId": "reset_confirm_api_v1_auth_reset_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetConfirm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set a new password using a reset link",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/reset/request": {
      "post": {
        "description": "**Always returns 202**, whether or not the account exists, has a verified chat, or delivery succeeded. Reporting any of those would be a reliable account-enumeration oracle (F8.AC10). Outcomes go to the audit log.",
        "operationId": "reset_request_api_v1_auth_reset_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request a reset link over Telegram",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/sessions/{session_id}": {
      "delete": {
        "operationId": "revoke_session_api_v1_auth_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke one of your sessions",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/telegram/verify/confirm": {
      "post": {
        "operationId": "telegram_verify_confirm_api_v1_auth_telegram_verify_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelegramVerifyConfirm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm the chat with the code it received",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/telegram/verify/start": {
      "post": {
        "description": "Proves the chat is reachable before it is trusted for password recovery. An unverified chat id means reset links go somewhere that may not be yours -- and you would only find out when you needed it.",
        "operationId": "telegram_verify_start_api_v1_auth_telegram_verify_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelegramVerifyStart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send a verification code to a Telegram chat",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}